requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
pymongo>=4.6.0
Pillow>=9.0.0
werkzeug>=2.3.0
opencv-python>=4.8.0
numpy>=1.23.0
tensorflow>=2.12.0
